Don't store last_processing_attempt if failure occurs at read stage
Intended to fix a problem when the hansard_process_sources management
command is being run from a dev box in the UK and the connection to
www.parliament.go.ke times out at the point the PDF file is requested
as this results in the source being silently skipped on the next run.
Apologies if this has unintended consequences in production.